moved aio_are_figure func to transmonee.py
