+++ dash_service/pages/child_protection.py
@@ -25,6 +25,7 @@ from dash_service.pages.transmonee import (
     filters,
     themes,
     aio_options,
+    aio_area_figure,
     fa,
     unicef_country_prog,
     programme_country_indexes,
@@ -234,6 +235,8 @@ page_config = {
     },
 }
 
+packed_config = {}
+
 register_page(
     __name__,
     # path_template="/transmonee/<page_slug>",
@@ -399,198 +402,15 @@ def set_default_compare(compare_options, selected_type, indicators_dict, theme):
     ],
     prevent_initial_call=True,
 )
-def aio_area_figure(
-    compare,
-    selections,
-    indicators_dict,
-    buttons_properties,
-    selected_type,
+def apply_aio_area_figure(
+    compare, selections, indicators_dict, buttons_properties, selected_type
 ):
-
-    # assumes indicator is not empty
-    indicator = [
-        but_prop["props"]["id"]["index"]
-        for but_prop in buttons_properties
-        if but_prop["props"]["active"] is True
-    ][0]
-
-    area = "AIO_AREA"
-    default_graph = indicators_dict[selections["theme"]][area].get(
-        "default_graph", "line"
-    )
-
-    fig_type = selected_type if selected_type else default_graph
-    fig_config = indicators_dict[selections["theme"]][area]["graphs"][fig_type].copy()
-    options = fig_config.get("options")
-    traces = fig_config.get("trace_options")
-    layout_opt = fig_config.get("layout_options")
-    dimension = False if fig_type in ["line", "map"] or compare == "TOTAL" else compare
-    indicator_name = str(indicator_names.get(indicator, ""))
-
-    data = get_filtered_dataset(
-        [indicator],
-        selections["years"],
-        selections["countries"],
+    return aio_area_figure(
         compare,
-        latest_data=False if fig_type == "line" else True,
-    )
-
-    # check if the dataframe is empty meaning no data to display as per the user's selection
-    if data.empty:
-        return EMPTY_CHART, "", [], [], []
-    else:
-        data.sort_values("OBS_VALUE", ascending=False, inplace=True)
-
-    # indicator card
-    card_config = [
-        elem
-        for elem in indicators_dict[selections["theme"]]["CARDS"]
-        if elem["indicator"] == indicator
-    ]
-
-    ind_card = (
-        []
-        if not card_config or "CARDS" not in indicators_dict[selections["theme"]]
-        else indicator_card(
-            selections,
-            card_config[0]["name"],
-            card_config[0]["indicator"],
-            card_config[0]["suffix"],
-            card_config[0].get("absolute"),
-            card_config[0].get("average"),
-            card_config[0].get("min_max"),
-            card_config[0].get("sex"),
-            card_config[0].get("age"),
-            page_prefix,
-        )
-    )
-
-    # lbassil: was UNIT_MEASURE
-    name = (
-        data[data["CODE"] == indicator]["Unit_name"].astype(str).unique()[0]
-        if len(data[data["CODE"] == indicator]["Unit_name"].astype(str).unique()) > 0
-        else ""
-    )
-    df_indicator_sources = df_sources[df_sources["Code"] == indicator]
-    unique_indicator_sources = df_indicator_sources["Source_Full"].unique()
-    source = (
-        "; ".join(list(unique_indicator_sources))
-        if len(unique_indicator_sources) > 0
-        else ""
-    )
-    source_link = (
-        df_indicator_sources["Source_Link"].unique()[0]
-        if len(unique_indicator_sources) > 0
-        else ""
-    )
-
-    options["labels"] = DEFAULT_LABELS.copy()
-    options["labels"]["OBS_VALUE"] = name
-
-    # set the chart title, wrap the text when the indicator name is too long
-    chart_title = textwrap.wrap(
-        indicator_name,
-        width=74,
-    )
-    chart_title = "<br>".join(chart_title)
-
-    # set the layout to center the chart title and change its font size and color
-    layout = go.Layout(
-        title=chart_title,
-        title_x=0.5,
-        font=dict(family="Arial", size=12),
-        legend=dict(x=1, y=0.5),
-        xaxis={
-            "categoryorder": "total descending",
-            "tickangle": -45,
-            "tickmode": "linear",
-            "tickfont_size": 10,
-        },
-    )
-    if layout_opt:
-        layout.update(layout_opt)
-
-    # Add this code to avoid having decimal year on the x-axis for time series charts
-    if fig_type == "line":
-        data.sort_values(by=["TIME_PERIOD", "Country_name"], inplace=True)
-        layout["xaxis"] = dict(
-            tickmode="linear",
-            tick0=selections["years"][0],
-            dtick=1,
-            categoryorder="total ascending",
-        )
-
-    if dimension:
-        # lbassil: use the dimension name instead of the code
-        dimension_name = str(dimension_names.get(dimension, ""))
-        options["color"] = dimension_name
-
-        # sort by the compare value to have the legend in the right ascending order
-        data.sort_values(by=[dimension], inplace=True)
-
-    # rename figure_type 'map': 'choropleth' (plotly express)
-    if fig_type == "map":
-        fig_type = "choropleth_mapbox"
-        options["range_color"] = [data.OBS_VALUE.min(), data.OBS_VALUE.max()]
-    fig = getattr(px, fig_type)(data, **options)
-    fig.update_layout(layout)
-    if traces:
-        fig.update_traces(**traces)
-
-    # countries not reporting
-    not_rep_count = np.setdiff1d(selections["count_names"], data.Country_name.unique())
-    # number of countries from selection
-    count_sel = len(selections["countries"])
-
-    return (
-        fig,
-        [
-            html.Div(
-                [
-                    html.P(
-                        "Source: ",
-                        style={
-                            "display": "inline-block",
-                            "textDecoration": "underline",
-                            "fontWeight": "bold",
-                        },
-                    ),
-                    html.A(
-                        f" {source}",
-                        href=source_link,
-                        target="_blank",
-                        id={
-                            "type": "area_sources",
-                            "index": f"{page_prefix}-AIO_AREA",
-                        },
-                        className="alert-link",
-                    ),
-                ],
-            )
-        ],
-        ind_card,
-        [
-            html.Div(
-                [
-                    html.P(
-                        "Countries without data: ",
-                        style={
-                            "display": "inline-block",
-                            "textDecoration": "underline",
-                            "fontWeight": "bold",
-                        },
-                    ),
-                    html.P(
-                        f" {len(not_rep_count)} / {count_sel}",
-                        style={
-                            "display": "inline-block",
-                            "fontWeight": "bold",
-                            "color": "#1cabe2",
-                            "whiteSpace": "pre",
-                        },
-                    ),
-                ]
-            )
-        ],
-        dcc.Markdown(["- " + "\n- ".join(sorted(not_rep_count, key=str.lower))]),
+        selections,
+        indicators_dict,
+        buttons_properties,
+        selected_type,
+        page_prefix,
+        packed_config,
     )
